% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{bind_parametrization.ExpFam_dist_ext}
\alias{bind_parametrization.ExpFam_dist_ext}
\title{ExpFam_dist with selected parametrisation}
\usage{
\method{bind_parametrization}{ExpFam_dist_ext}(dist.obj, param.name,
  num.opt = TRUE)
}
\arguments{
\item{dist.obj}{object of the class \code{\link{ExpFam_dist_ext}}}

\item{param.name}{(\code{character}) the name of the parametrisation used as a theta}

\item{num.opt}{(\code{logical}) should numerically-optimized functions be used in the defined function}
}
\value{
object of class \code{\link{ExpFam_dist}}
}
\description{
ExpFam_dist with selected parametrisation
}

