% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{bind_parametrization}
\alias{bind_parametrization}
\title{The generic function for selecting a parametrization}
\usage{
bind_parametrization(x, ...)
}
\description{
The generic function for selecting a parametrization
}

