% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ext_instances.R
\name{dist_ext_Bernoulli}
\alias{dist_ext_Bernoulli}
\title{Bernoulli distribution}
\usage{
dist_ext_Bernoulli()
}
\value{
object of class \code{\link{ExpFam_dist_ext}}
}
\description{
Bernoulli distribution
}

