% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ext_instances.R
\name{dist_ext_Exponential}
\alias{dist_ext_Exponential}
\title{Exponential distribution}
\usage{
dist_ext_Exponential()
}
\value{
object of class \code{\link{ExpFam_dist_ext}}
}
\description{
Exponential distribution
}

