% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_bind_parametrization.ExpFam_dist_ext}
\alias{ExpFam_bind_parametrization.ExpFam_dist_ext}
\title{ExpFam_dist with selected parametrisation}
\usage{
\method{ExpFam_bind_parametrization}{ExpFam_dist_ext}(dist.obj, param.name,
  num.opt = TRUE)
}
\arguments{
\item{dist.obj}{object of the class \code{\link{ExpFam_dist_ext}}}

\item{param.name}{(\code{character}) the name of the parametrisation used as a theta}
}
\description{
ExpFam_dist with selected parametrisation
}

