% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_bind_parametrization}
\alias{ExpFam_bind_parametrization}
\title{The generic function for selecting a parametrization}
\usage{
ExpFam_bind_parametrization(x, ...)
}
\description{
The generic function for selecting a parametrization
}

