% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_density.ExpFam_dist_canonical}
\alias{ExpFam_density.ExpFam_dist_canonical}
\title{The density function for canonical parametrization}
\usage{
\method{ExpFam_density}{ExpFam_dist_canonical}(org.dist, eta, num.opt = TRUE,
  verify = TRUE)
}
\description{
\eqn{d(x|\eta) = h(x) \exp{(\eta^T S(x) - A(\eta))}}
}

