% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_density.ExpFam_param}
\alias{ExpFam_density.ExpFam_param}
\title{The density function for given parametrization}
\usage{
\method{ExpFam_density}{ExpFam_param}(par.dist, theta, num.opt = TRUE,
  verify = TRUE)
}
\description{
\eqn{d(x|\eta) = h(x) \exp{(\eta^T S(x) - A(\eta))}}
}

