% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_density}
\alias{ExpFam_density}
\title{The generic density function for canonical parametrization}
\usage{
ExpFam_density(x, ...)
}
\description{
The generic density function for canonical parametrization
}

