% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{ExpFam_density_eta.ExpFam_dist}
\alias{ExpFam_density_eta.ExpFam_dist}
\title{The density function for mean parametrization}
\usage{
\method{ExpFam_density_eta}{ExpFam_dist}(org.dist, eta, num.opt = TRUE,
  verify = TRUE)
}
\description{
\eqn{d(x|\eta) = h(x) \exp{(\eta^T S(x) - A(\eta))}}
}

