% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{ExpFam_density_eta}
\alias{ExpFam_density_eta}
\title{The generic density function for mean parametrization}
\usage{
ExpFam_density_eta(x, ...)
}
\description{
The generic density function for mean parametrization
}

