% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{ExpFam_density_theta.ExpFam_dist}
\alias{ExpFam_density_theta.ExpFam_dist}
\title{The density function for canonical parametrization}
\usage{
\method{ExpFam_density_theta}{ExpFam_dist}(org.dist, theta)
}
\description{
The density function for canonical parametrization
}

