% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{ExpFam_density_theta}
\alias{ExpFam_density_theta}
\title{The generic density function for canonical parametrization}
\usage{
ExpFam_density_theta(x, ...)
}
\description{
The generic density function for canonical parametrization
}

