% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_dist_canonical}
\alias{ExpFam_dist_canonical}
\title{Components of Exponential Family distribution}
\usage{
ExpFam_dist_canonical(h, A.from.eta, S, eta.dim, A.grad, A.hess, eta.in.domain)
}
\arguments{
\item{h}{(function) the measure}

\item{A.from.eta}{(function) value of log-normalizer for natural parameter in canonical parametrization}

\item{S}{(function) vector of sufficient statistics}

\item{eta.dim}{(numeric), number of parameters of the exponential family being defined}

\item{A.grad}{(function) gradient of the function A.from.eta}

\item{A.hess}{(function) hessian of the function A.from.eta}
}
\description{
Components of Exponential Family distribution
}

