% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_dist_ext}
\alias{ExpFam_dist_ext}
\title{ExpFam distribution}
\usage{
ExpFam_dist_ext(canonical.param, parametrisations)
}
\arguments{
\item{canonical.param}{canonical parametrisation, object of class \code{\link{ExpFam_dist_canonical}}}

\item{parametrisations}{list of parametrisations, objects of class \code{\link{ExpFam_param}}}
}
\description{
The class is a container for the canonical form and various parametrisations
}

