% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparams.R
\name{ExpFam_dist_reparam}
\alias{ExpFam_dist_reparam}
\title{Exponential Family Reparametrization}
\usage{
ExpFam_dist_reparam(y.from.x, x.from.y, y.grad, x.grad, y.hess, x.hess,
  x.in.domain, y.in.domain)
}
\arguments{
\item{y.from.x}{(\code{function}) y = f(x)}

\item{x.from.y}{(\code{function}) x = f^{-1}(y) = finv(y)}

\item{y.grad}{(\code{function}) f'(x)}

\item{x.grad}{(\code{function}) finv'(y)}

\item{y.hess}{(\code{function}) f''(x)}

\item{x.hess}{(\code{function}) finv''(y)}

\item{x.in.domain}{(\code{function})}

\item{y.in.domain}{(\code{function}) returns \code{TRUE} iff all}
}
\description{
The class describing 1-1 mapping of the parameters of the exponential family,
consisting of:
\itemize{
  \item{the mapping and inverse mapping,}
  \item{gradients,}
  \item{Hessians,}
  \item{domains characteristic functions}
}
All components are assumed to be vectorisable,
when dimension > 1 the mapped points are assumed to be passed as rows of the matrix.
}

