% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_param}
\alias{ExpFam_param}
\title{Parametrization of the canonical ExpFam distribution}
\usage{
ExpFam_param(org.dist, eta.from.theta, theta.from.eta, B.from.theta, B.grad,
  B.hess, theta.in.domain, param.type)
}
\arguments{
\item{org.dist}{(object of class \code{\link{ExpFam_dist_canonical}}) the distribution being parametrized}

\item{eta.from.theta}{(\code{function}) mapping to canonical parameters \eqn{\eta(\theta)}}

\item{theta.from.eta}{(\code{function}) mapping from the canonical parameters \eqn{\theta(\eta)}}

\item{B.from.theta}{(\code{function}) the reparametrized log-normalizer of the distribution,
simplified and numerically sound \eqn{B(\theta) = A(\eta(\theta))}}

\item{B.grad}{(\code{function}) the gradient of the reparametrized log-normalizer of the distribution,
simplified and numerically sound \eqn{B'(\theta) = A'(\eta(\theta)) \eta'(\theta)}}

\item{B.hess}{(\code{function}) the Hessian of the reparametrized log-normalizer of the distribution,
simplified and numerically sound
\eqn{B''(\theta) = A''(\eta(\theta)) (\eta'(\theta))^2 + A'(\eta(\theta)) \eta''(\theta) }}

\item{theta.in.domain}{(\code{function}) characteristic function}

\item{param.type}{(\code{character}) the parametrisation "identifier" (e.g. "Bernoulli, mean")}
}
\description{
The class provides a description of the mapping \eqn{\eta = f(\theta)},
where \eqn{\eta} is a canonical parameter.
}

