% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_reparametrize.ExpFam_dist_canonical}
\alias{ExpFam_reparametrize.ExpFam_dist_canonical}
\title{Reparametrization of the ExpFam distribution in canonical form}
\usage{
\method{ExpFam_reparametrize}{ExpFam_dist_canonical}(param.obj, reparam.obj,
  hints = list(), param.type)
}
\description{
Reparametrization of the ExpFam distribution in canonical form
}
\examples{
  ExpFam_reparametrize.ExpFam_dist_canonical(
   param.obj = ,
   reparam.obj = Reparam_Logit(),
   hints = list(
       B.from.theta = function(theta) -log(1 - theta), #log-normaliser for mean parametrisation
       B.grad = function(theta) -1 / (1 + theta),
       B.hess = function(theta) 1 / (1 + theta) ^ 2
      ),
   param.type = "mean"
  )

}

