% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_param_objects.R
\name{ExpFam_reparametrize.ExpFam_param}
\alias{ExpFam_reparametrize.ExpFam_param}
\title{Exponential Family Reparametrisation}
\usage{
\method{ExpFam_reparametrize}{ExpFam_param}(param.obj, reparam.obj, param.type)
}
\arguments{
\item{param.obj}{object of class \code{\link{ExpFam_param}}}

\item{reparam.obj}{reparametrisation descriptor,
object of class \code{\link{ExpFam_dist_reparam}}}

\item{param.type}{the name for resulting parametrisation}
}
\description{
Exponential Family Reparametrisation
}

