% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{has_stat_GLM}
\alias{has_stat_GLM}
\title{The generic property of having a GLM-compatible object}
\usage{
has_stat_GLM(x, ...)
}
\description{
The generic property of having a GLM-compatible object
}

