% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{reparametrize_function_grad_with_proto}
\alias{reparametrize_function_grad_with_proto}
\title{Gradient of function reparametrization}
\usage{
reparametrize_function_grad_with_proto(f.from.z = NULL, f.from.z.grad,
  z.from.x, z.from.x.grad, res.proto, simplify = TRUE)
}
\arguments{
\item{f.from.z}{(\code{function}) f(z)}

\item{f.from.z.grad}{(\code{function}) df/dz (z)}

\item{z.from.x}{(\code{function}) z(x)}

\item{z.from.x.grad}{(\code{function}) dz/dx (x)}

\item{res.proto}{"prototype" function with desired formal name, e.g. \code{function(x) NULL}}

\item{simplify}{(\code{logical}) should the expression simplification be used?}
}
\description{
Function calculates the gradient df(z(x))/dx of f(z(x))
}

