% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{reparametrize_function_hess_with_proto}
\alias{reparametrize_function_hess_with_proto}
\title{Hessian of function reparametrization}
\usage{
reparametrize_function_hess_with_proto(f.from.z = NULL, f.from.z.grad,
  f.from.z.hess, z.from.x, z.from.x.grad, z.from.x.hess, res.proto,
  simplify = TRUE)
}
\arguments{
\item{f.from.z}{(\code{function}) f(z)}

\item{f.from.z.grad}{(\code{function}) df/dz (z)}

\item{f.from.z.hess}{(\code{function}) d^2 f/dz^2 (z)}

\item{z.from.x}{(\code{function}) z(x)}

\item{z.from.x.grad}{(\code{function}) dz/dx (x)}

\item{z.from.x.hess}{(\code{function}) d^2 z/dx^2 (x)}

\item{res.proto}{"prototype" function with desired formal name, e.g. \code{function(x) NULL}}

\item{simplify}{(\code{logical}) should the expression simplification be used?}
}
\description{
Function calculates the hessian d{^2}f(z(x))/dx^2 of f(z(x))
}

