% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_objects.R
\name{ExpFam_dist}
\alias{ExpFam_dist}
\title{General components of Exponential Family distribution}
\usage{
ExpFam_dist(h, eta.from.theta, theta.from.eta, B.from.theta, A.from.eta, S)
}
\arguments{
\item{h}{(function) the measure}

\item{eta.from.theta}{(function) canonical parameter from mean parameter}

\item{theta.from.eta}{(function) mean parameter from canonical parameter}

\item{B.from.theta}{(function) value of log-normalizer for mean parameter}

\item{A.from.eta}{(function) value of log-normalizer for natural parameter in canonical parametrization}

\item{S}{(function) vector of sufficient statistics}
}
\description{
General components of Exponential Family distribution
}

