% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{reparametrize_function_with_proto}
\alias{reparametrize_function_with_proto}
\title{Function reparametrization}
\usage{
reparametrize_function_with_proto(f1, f2, res.proto)
}
\arguments{
\item{f1}{function 1}

\item{f2}{function 2}

\item{res.proto}{resulting function prototype with NULL body, e.g. \code{function(theta) NULL}}
}
\description{
Creates a function being a reparametrisation (with simplification) of
 \code{f1 * f2}
}
\examples{

 reparametrize_function_with_proto(
   function(x) x^2,
   function(y) y^3,
   function(z) NULL
   )
   # function(z) z^6

 reparametrize_function_with_proto(
   function(x) x^2 + 2*x,
   function(y) (y+1)^3,
   function(z) NULL
   )
   # function(z) ((1 + z)^3 + 2) * (1 + z)^3

}

